Chdir on endpoint.go init
for all imports of enpoint.go (init()) ensure that we change WORKING DIRECTORY (os.Chdir) to directory of our bin.  From a probe this will be the VST.
All writes are now relative to the VST folder.
PWD is also updated to not calculate bin folder, but use what we've already calculated on init.
This ensures we run from the "bindir" even if Endpoint.Start() isn't used.
The net effect here is that our tests will always write to the .vst folder